--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      { "useBuiltIns": "usage", "corejs": 3 }
-    ]
-  ]
-}
