+++ .eslintrc
@@ -1,12 +1,3 @@
 {
-  "parser": "babel-eslint",
-  "extends": "airbnb/base",
-  "env": {
-    "mocha": true,
-    "node": true
-  },
-  "rules": {
-    "comma-dangle": 0,
-    "no-console": 0
-  }
+  "extends": "@terascope"
 }
