+++ .gitignore
@@ -2,5 +2,6 @@
 *.log
 node_modules
 lib
+dist
 tmp
 coverage
