--- .npmignore
@@ -1,7 +1,0 @@
-.DS_Store
-*.log
-src
-test
-tmp
-coverage
-.vscode
