+++ README.md
@@ -66,7 +66,7 @@ yarn add @terascope/fetch-github-release
 ### Usage
 
 ```javascript
-const downloadRelease = require('@terascope/fetch-github-release');
+const { downloadRelease } = require('@terascope/fetch-github-release');
 
 const user = 'some user';
 const repo = 'some repo';
