+++ bin/fetch-github-release
@@ -2,4 +2,4 @@
 
 var path = require('path');
 var fs = require('fs');
-require(path.join(path.dirname(fs.realpathSync(__filename)), '../lib/cli'));
+require(path.join(path.dirname(fs.realpathSync(__filename)), '../dist/src/cli'));
