+++ jest.config.js
@@ -1,0 +1,30 @@
+'use strict';
+
+module.exports = {
+    verbose: true,
+    testEnvironment: 'node',
+    setupFilesAfterEnv: ['jest-extended'],
+    collectCoverage: true,
+    coverageReporters: ['json', 'lcov', 'text', 'html'],
+    coverageDirectory: 'coverage',
+    collectCoverageFrom: [
+        '<rootDir>/src/**/*.ts',
+        '!<rootDir>/**/coverage/**',
+        '!<rootDir>/**/*.d.ts',
+        '!<rootDir>/**/dist/**',
+        '!<rootDir>/**/coverage/**'
+    ],
+    testMatch: [
+        '<rootDir>/test/**/*-spec.{ts,js}',
+        '<rootDir>/test/*-spec.{ts,js}',
+    ],
+    preset: 'ts-jest',
+    globals: {
+        'ts-jest': {
+            tsconfig: './tsconfig.json',
+            diagnostics: true,
+        },
+        ignoreDirectories: ['dist'],
+        availableExtensions: ['.js', '.ts']
+    }
+};
