--- src/index.js
@@ -1,3 +1,0 @@
-import downloadRelease from './downloadRelease';
-
-module.exports = downloadRelease;
