+++ src/index.ts
@@ -1,0 +1,3 @@
+import { downloadRelease } from './downloadRelease';
+
+export { downloadRelease };
