+++ test/index-spec.ts
@@ -1,0 +1,8 @@
+import 'jest-extended';
+import { downloadRelease } from '../src';
+
+describe('fetch-github-release', () => {
+    it('should expose a function', () => {
+        expect(downloadRelease).toBeFunction();
+    });
+});
