--- test/index.js
@@ -1,7 +1,0 @@
-import downloadRelease from '../src';
-
-describe('fetch-github-release', () => {
-  it('should expose a function', () => {
-    downloadRelease.should.be.a.Function();
-  });
-});
