+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+        "rootDir": ".",
+        "outDir": "dist",
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "target": "es2019",
+        "skipLibCheck": true,
+        "experimentalDecorators": true,
+        "strict": true,
+        "sourceMap": true,
+        "inlineSourceMap": false,
+        "noImplicitReturns": true,
+        "preserveConstEnums": true,
+        "esModuleInterop": true,
+        "resolveJsonModule": true,
+        "suppressImplicitAnyIndexErrors": true
+    },
+    "include": [
+        "src/**/*",
+        "test/**/*"
+    ]
+}
